requirements.txt
opencv-python
numpy
Pillow
tqdm
scikit-image
noise
opensimplex
perlin-noise
